streamlit_app/requirements.txt
streamlit>=1.28.0
h5py>=3.10.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
